add redeem script sanity check
When collaborating to build a multisig address, the user will want to be
absolutely certain that the p2sh address is correct, because they don't want to
accidentally send bitcoins into obvlivion.  Each user will want to be confident
that their public key is in the redeem script. This is part of that check. The
user can see that their public key is in the list, and this check is making
sure that the redeem script actually contains that public key. Another check
that I will implement shortly is that the user will want to be confident that
they actually have the privatate key corresponding to their public key.