added ability to check which priv keys a user has
It will be imortant for users to be able to confirm that they actually have the
private key corresponding to a public key in the p2sh redeem script. That is
what this commit does. Users can show a list of pubkeys in the db, and show
which pubkeys they have the privkey for. Checks are run both to make sure the
corresponding pubkey is in the user's wallet, AND that the corresponding
private key can actually rederive the public key correctly.