add validate.inputtxs
The purpose of this validator is to make sure tha the inputtxs, which are
normally retrieved from bitcoind but may also be passed around in this file,
actually correspond to the "inputs" as listed in the db (but not the inputs in
the raw tx). This to make sure that the inputtxs had been retrieved correctly.
Two checks are performed: 1) that each input is present as an inputtxs, and
that the tx in the inputtx hashes to the txid correctly (i.e. that the correct
tx is there).