add validate.all convenience command
The purpose of validate.all is to run all validate functions so that someone
who is interested in signing a transaction can confirm at a glance that it is
something that they want to sign. There is only one validate function that is
yet to be written that should be added (that I can think of), and that is a
validate.sigs validator that checks that all existing signatures are valid, and
counts then, and sees if the transaction can be broadcast. Actually, one more
after that should be something like validate.standard, to confirm that the
transaction is standard before broadcasting it (so that it will actually be
included in a block).