add validate.p2sh
One of the most important ways things can go wrong is for the p2sh address to
be generated incorrectly, leaving any bitcoins sent to that address
unretrievable. This validate function recalculates the p2sh address
corresponding to the redeemScript, and confirms that the wallet contains this
address and script. This validation combined with validate.redeem (ensuring the
redeem script is what you think it is) and validate.pubkeys (ensuring that you
have the private key for at least one of the public keys) makes it unlikely
that you have generated an address from which bitcoins can't be retrieved