added validate.sigs
The purpose of this function is to loop through the inputs and make sure any
sigs are actually valid. This requires a bit of intelligence to figure out what
is actually a signature and what is something else. We assume if you are runnin
this function that the transaction was created with cosign and that there is
only one input (this will need to be extended later but works fine for our
purposes for now). It will count the number of signatures and explain whether
the transaction is fully signed and ready to be broadcast or not. For the
moment, the signatures are not checked for validity - it is assumed they are
valid. If not, the transaction will not be included in a block. The signatures
probably should be checked for validity, but it is not imperative - this
function for now is mostly about informing the user about how many more
signatures are needed.